% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{ASER}
\alias{ASER}
\title{A illustration of ASE estimate}
\usage{
ASER(UX, Uy, noo)
}
\arguments{
\item{UX}{a list of all covariate samples observe}

\item{Uy}{a list of all responses observe}

\item{noo}{the size of initial sample size}
}
\value{
a list contain the estimate of coefficient and sample size used
}
\description{
ASE is a variable select method
}
\examples{
\dontrun{
result<-ASER(UX,Uy,40)
}
}
