% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{data}
\alias{data}
\title{A illustration dataset}
\description{
A dataset used to illustrate the performance of ASER and MQLE.
}
\examples{
\dontrun{
n=3000
set.seed(1234)
pd=36
beta0=c(1,-1,1.5,-2,rep(0,pd))
p<-length(beta0)
C=diag(length(beta0))
for (i in 1:length(beta0)) {
 for (j in 1:length(beta0)) {
   if(i==j)
     C[i,j]=0.5
   else
     C[i,j]=0.2
 }
}
n=3000
UX=list()
for(i in 1:n){
 UX[[i]]=mvrnorm(n=5,rep(0,length(beta0)), C)
}
Uy=list()
data(data)
for (i in 1:n) {
 Uy[[i]]<-data[,i]
}
result1<-ASER(UX,Uy,noo=50)
result2<-MQLE(UX,Uy,noo=50)
}
}
