% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{Epan}
\alias{Epan}
\title{The hw1 Question4}
\usage{
Epan(n)
}
\arguments{
\item{n}{the sample size}
}
\value{
a random sample of size \code{n}
}
\description{
A function to generate a random sample of size \code{n} from the the pdf is rescaled Epanechnikov kernel
}
\examples{
\dontrun{
set.seed(1234)
y<-Epan(1e4)
print(y)
}
}
