% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{f_MLE}
\alias{f_MLE}
\title{MLE and EM compare}
\usage{
f_MLE(lamda)
}
\arguments{
\item{lamda}{the true value}
}
\value{
the compare of  EM and MLE
}
\description{
A illustration of the MLE and EM method
}
\examples{
\dontrun{
set.seed(12345)
U<-c(11,8,27,13,16,0,23,10,24,2)
V<-U+1
sol1<-uniroot(f_MLE,c(0,10))
lamda1<-round(sol1$root,6)
lamdan<-0.08
for (i in 1:100) {
  lamdan<-lamdan
  sum2<-numeric(10)
  for (i in 1:10) {
    fenzi1=(U[i]-(V[i]+1/lamdan)*exp(lamdan*(U[i]-V[i]))+1/lamdan)
    fenmu1=1-exp(-lamdan*(V[i]-U[i]))
    sum2[i]=fenzi1/fenmu1
  }
  lamdan<-1/mean(sum2)
}
lamda2<-round(lamdan,6)
c(lamda1,lamda2)
}
}
