% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{gbR}
\alias{gbR}
\title{The hw7 Question3}
\usage{
gbR(N, burn)
}
\arguments{
\item{N}{the length of whole chain}

\item{burn}{the length of pre-burn}
}
\value{
a MC chain
}
\description{
A illustration of Gibbs sampler
}
\examples{
\dontrun{
set.seed(1234)
result<-gbR(1e4,1e3)
}
}
