% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GibbsC}
\alias{GibbsC}
\title{A Gibbs sampler using Rcpp}
\usage{
GibbsC(N, thin, a, b, n)
}
\arguments{
\item{N}{the number of samples}

\item{thin}{the number of between-sample random numbers}

\item{a}{the first parameter of Gibbs sampler proposition distribution of rbeta of y}

\item{b}{the second parameter of Gibbs sampler proposition distribution of rbeta of y}

\item{n}{the parameter of Gibbs sampler proposition distribution of rbinom of x}
}
\value{
a random sample of size \code{N}
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
rnC <- GibbsC(1e3,10,2,3,10)
par(mfrow=c(2,1));
plot(rnC[,1],type='l')
plot(rnC[,2],type='l')
}
}
