% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{hw2Q1}
\alias{hw2Q1}
\title{The hw2 Question1}
\description{
What value should take to minimize the asymptotic variance of puffen?
}
\examples{
\dontrun{
rho<-c(0.5,0.8,1)
set.seed(1234)
va_pi<-rep(0,length(rho))
K<-100
pi_hat<-rep(0,K)
for (i in 1:3) {
 for (k in 1:100) {
   d<-1
   l<-rho[i]*d
 n_times<-1e6
   x_i<-runif(n_times,0,d/2)
   y_i<-runif(n_times,0,pi/2)
   pi_hat[k]<-2*l/d/mean(l/2*sin(y_i)>x_i)
 }
 va_pi[i]<-var(pi_hat)
}
paste("the variance in different rho(0.5,0.8,1.0) is:",va_pi[1],";",va_pi[2],";",va_pi[3])
}
}
