% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{hw2Q2}
\alias{hw2Q2}
\title{The hw2 Question2}
\description{
A illustration of antithetic variate approach of Monte Carlo integration
}
\examples{
\dontrun{
result1<-1/2*(1/2*(exp(2)-1)-(exp(1)-1)^2)
result2<- -3/4*exp(2)+5/2*exp(1)-5/4
print((result1-result2)/result1)
}
}
