% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{hw2Q3}
\alias{hw2Q3}
\title{The hw2 Question3}
\description{
A illustration of compare of the antithetic variate approach and the simple Monte Carlo method
}
\examples{
\dontrun{
n_times<-1e5
set.seed(1234)
mcsimple<-rep(0,1000)
mcanti<-rep(0,1000)
for (k in 1:1000) {
 mcsimple[k]<-mean(exp(runif(n_times)))
 U_k<-runif(n_times/2)
 one_minus_U_k<-1-U_k
 mcanti[k]<-mean((exp(U_k)+exp(one_minus_U_k))/2)
 }
paste(round(mean(mcsimple),5),round(mean(mcanti),5),round(exp(1)-1,5))
paste(round((var(mcsimple)-var(mcanti))/var(mcsimple),6))
}
}
