% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{hw3}
\alias{hw3}
\title{The hw3}
\description{
A illustration of importance sampling
}
\examples{
\dontrun{
m<-1e6
set.seed(123)
est<-sd<-numeric(2)
g<-function(x)x^2*exp(-x^2/2)/sqrt(2*pi)*(x>1)
true_value<-integrate(g,1,Inf)$value
x1<-sqrt(rchisq(m,1))+1
f_1<-2*dnorm(x1,1,1)
f_g_1<-g(x1)/f_1
est[1]<-mean(f_g_1)
sd[1]<-sd(f_g_1)
x2<-rgamma(m,3/2,2)+1
f_2<-dgamma(x2-1,3/2,2)
f_g_2<-g(x2)/f_2
est[2]<-mean(f_g_2)
sd[2]<-sd(f_g_2)
}
}
