% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{hw4Q1}
\alias{hw4Q1}
\title{The hw4 Question1}
\description{
A illustration compare the p-value adjust methods
}
\examples{
\dontrun{
M<-m<-1000
l1<-0.95*m
alpha<-0.1
FWER_bon<-FWER_bh<-FDR_bon<-FDR_bh<-TPR_bon<-TPR_bh<-numeric(M)
set.seed(123)
for (i in 1:M) {
  p1<-runif(l1)
  p2<-rbeta(0.05*m,0.1,1)
  p<-c(p1,p2)
  p_bon<-p.adjust(p,method='bonferroni')
  p_bh<-p.adjust(p,method='fdr')
  FWER_bon[i]<-1-all(p_bon[1:l1]>=alpha)
  FWER_bh[i]<-1-all(p_bh[1:l1]>=alpha)
  FDR_bon[i]<-sum(p_bon[1:l1]<=alpha)/sum(p_bon<=alpha)
  FDR_bh[i]<-sum(p_bh[1:l1]<=alpha)/sum(p_bh<=alpha)
  TPR_bon[i]<-mean(p_bon[l1:m]<=alpha)
  TPR_bh[i]<-mean(p_bh[l1:m]<=alpha)
}
FW_bon<-mean(FWER_bon)
FW_bh<-mean(FWER_bh)
FD_bon<-mean(FDR_bon)
FD_bh<-mean(FDR_bh)
TPRbon<-mean(TPR_bon)
TPRbh<-mean(TPR_bh)
re<-data.frame(FWER=round(c(FW_bon,FW_bh),4),
               FDR=round(c(FD_bon,FD_bh),4),
               TPR=round(c(TPRbon,TPRbh),4),row.names=c("bonf","b-h"))
re 
}
}
