% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{hw4Q2}
\alias{hw4Q2}
\title{The hw4 Question2}
\description{
A illustration of bootstrap estimate of bias and sd
}
\examples{
\dontrun{
lamda<-2
n<-c(5,10,20)
B<-1e3
m<-1e3
set.seed(123)
lamdaf<-function(x,i) 1/mean(x[i])
bias_the<-sd_the<-bias_boot<-sd_boot<-numeric(length(n))
bias_bootm<-sd_bootm<-numeric(m)
for (i in 1:3) {
  bias_the[i]<-lamda/(n[i]-1)
  sd_the[i]<-lamda*n[i]/((n[i]-1)*sqrt(n[i]-2))
  for (j in 1:m) {
    data<-rexp(n[i],lamda)
    om<-boot(data,lamdaf,B)
    bias_bootm[j]<-mean(om$t)-om$t0
    sd_bootm[j]<-sd(om$t)
  }
  bias_boot[i]<-mean(bias_bootm)
  sd_boot[i]<-mean(sd_bootm)
}
re<-data.frame(bias_boot=round(c(bias_boot[1],bias_boot[2],bias_boot[3]),4),
               bias_theory=round(c(bias_the[1],bias_the[2],bias_the[3]),4),
               sd_boot=round(c(sd_boot[1],sd_boot[2],sd_boot[3]),4),
               sd_theory=round(c(sd_the[1],sd_the[2],sd_the[3]),4),
               row.names=c("n=5","n=10","n=20"))
re 
}
}
