% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{hw4Q3}
\alias{hw4Q3}
\title{The hw4 Question3}
\description{
A illustration of the bootstrap t confidence interval estimate for correlation statistic
}
\examples{
\dontrun{
set.seed(123)
m<-10
B<-1e2
n<-nrow(law)
LCL<-UCL<-numeric(m)
boot.cor<-function(x,i) cor(x[i,1],x[i,2])
corr<-boot.cor(law,1:n)
corr_b<-t<-numeric(B)
index<-numeric(n)
for (i in 1:m) {
  for (b in 1:B) {
    index<-sample(1:n,replace=TRUE)
    corr_b[b]<-boot.cor(law,index)
    ob<-boot(law[index,],boot.cor,B)
    t[b]<-(corr_b[b]-corr)/sd(ob$t)
  }
  LCL[i]<-corr-quantile(t,0.975)*sd(corr_b)
  UCL[i]<-corr-quantile(t,0.025)*sd(corr_b)
}
cat("the 95\% CI is:","[",round(mean(LCL),4),",",round(mean(UCL),4),"]")
}
}
