% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{hw5Q2}
\alias{hw5Q2}
\title{The hw5 Question2}
\description{
A illustration of the jackknife estimates of bias and standard error
}
\examples{
\dontrun{
attach(scor)
data2<-as.matrix(scor)
n<-nrow(data2)
theta_jacm<-numeric(n)
lambda<-eigen(cov(data2))$values
theta_hat<-max(lambda)/sum(lambda)
for (i in 1:n) {
  lambda_n<-eigen(cov(data2[-i,]))$values
  theta_jacm[i]<-max(lambda_n)/sum(lambda_n)
}
bias_jacm<-(n-1)*(mean(theta_jacm)-theta_hat)
se_jacm<-sqrt((n-1)/n*sum((theta_jacm-mean(theta_jacm))^2))
cat('bias','=',round(bias_jacm,4),'sd','=',round(se_jacm,4))
}
}
