% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{hw5Q3}
\alias{hw5Q3}
\title{The hw5 Question3}
\description{
A illustration of the leave-two-out cross validation to compare the models
}
\examples{
\dontrun{
attach(ironslag)
n<-length(magnetic)
N<-n*(n-1)/2
m<-1
err1<-err2<-err3<-err4<-numeric(N)
for (i in 1:(n-1)) {
  for (j in (i+1):n){
    k<-c(i,j)
    y<-magnetic[-k]
    x<-chemical[-k]
    Mod1<-lm(y~x)
    yhat1<-Mod1$coef[1]+Mod1$coef[2]*chemical[k]
    err1[m]<-sum((magnetic[k]-yhat1)^2)
    Mod2<-lm(y~x+I(x^2))
    yhat2<-Mod2$coef[1]+Mod2$coef[2]*chemical[k]+Mod2$coef[3]*chemical[k]^2
    err2[m]<-sum((magnetic[k]-yhat2)^2)
    Mod3<-lm(log(y)~x)
    logyhat3<-Mod3$coef[1]+Mod3$coef[2]*chemical[k]
    yhat3<-exp(logyhat3)
    err3[m]<-sum((magnetic[k]-yhat3)^2)
    Mod4<-lm(log(y)~log(x))
    logyhat4<-Mod4$coef[1]+Mod4$coef[2]*log(chemical[k])
    yhat4<-exp(logyhat4)
    err4[m]<-sum((magnetic[k]-yhat4)^2)
    m<-m+1
  }
}
cat('the errors are:',round(c(mean(err1),mean(err2),mean(err3),mean(err4)),4))
}
}
