% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{hw9Q6}
\alias{hw9Q6}
\title{The hw9 Question6}
\description{
Use R package \code{microbenchmark} to compare the performance of R function \code{GibbsR} and C function \code{GibbsC}.
}
\examples{
\dontrun{
tm1 <- microbenchmark(
  rnR = GibbsR(1e3,10,2,3,10),
  rnC = GibbsC(1e3,10,2,3,10)
)
summary(tm1)[,c(1,3,5,6)]
}
}
