% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inverseMatrix}
\alias{inverseMatrix}
\title{Get the inverse of a matrix using Rcpp}
\usage{
inverseMatrix(mat)
}
\arguments{
\item{mat}{A matrix which has inverse}
}
\value{
the inverse of mat
}
\description{
Compute inverse using Rcpp
}
\examples{
\dontrun{
inverseMatrix(mat)
}
}
