% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{invlap}
\alias{invlap}
\title{Generate random sample from the standard Laplace distribution}
\usage{
invlap(n)
}
\arguments{
\item{n}{the sample size}
}
\value{
a random sample of size \code{n}
}
\description{
Use the inverse transform method to generate a random sample of size \code{n} from the standard Laplace distribution
}
\examples{
\dontrun{
set.seed(1234)
y<-invlap(1000)
print(y)
}
}
