% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{microben}
\alias{microben}
\title{Use microbenchmark compare the performance of R and Rcpp}
\description{
Use R package microbenchmark to compare the performance of R function \code{GibbsR} and C function \code{GibbsC}.
}
\examples{
\dontrun{
tm1 <- microbenchmark(
  rnR = GibbsR(1e3,10,2,3,10),
  rnC = GibbsC(1e3,10,2,3,10)
)
summary(tm1)[,c(1,3,5,6)]
}
}
