% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{MQLE}
\alias{MQLE}
\title{A illustration of MQLE estimate}
\usage{
MQLE(UX, Uy, noo)
}
\arguments{
\item{UX}{a list of all covariate samples observe}

\item{Uy}{a list of all responses observe}

\item{noo}{the size of initial sample size}
}
\value{
a list contain the estimate of coefficient and sample size used
}
\description{
MQLE is a coefficient estimate method
}
\examples{
\dontrun{
result<-MQLE(UX,Uy,40)
}
}
