% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{mybeta}
\alias{mybeta}
\title{Generate random sample from the beta(a,b) distribution}
\usage{
mybeta(n, a, b)
}
\arguments{
\item{n}{the sample size}

\item{a}{the Beta param1}

\item{b}{the Beta param2}
}
\value{
a random sample of size \code{n}
}
\description{
A function to generate a random sample of size \code{n} from the Beta(\code{a},\code{b}) distribution by the acceptance-rejection method
}
\examples{
\dontrun{
set.seed(1234)
y<-mybeta(1000,3,2)
print(y)
}
}
