% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{mycvtest}
\alias{mycvtest}
\title{Two-sample Cramer-von Mises test}
\usage{
mycvtest(x, y, R)
}
\arguments{
\item{x}{a vector of type 1 samples}

\item{y}{a vector of type 2 samples}

\item{R}{permutation times}
}
\value{
a p-value
}
\description{
The two-sample Cramer-von Mises test for equal distributions as a permutation test
}
\examples{
\dontrun{
attach(chickwts)
x<-sort(as.vector(weight[feed=="soybean"]))
y<-sort(as.vector(weight[feed=="linseed"]))
detach(chickwts)
R<-1e4
mycvtest(x,y,R)
}
}
