% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{mysample}
\alias{mysample}
\title{Reproduce the function Sample}
\usage{
mysample(x, size, pro)
}
\arguments{
\item{x}{a vector of one or more elements from which to choose, or a positive integer}

\item{size}{a non-negative integer giving the number of items to choose}

\item{pro}{description a vector of probability weights for obtaining the elements of the vector being sampled}
}
\value{
a random sample of size \code{size}
}
\description{
Use the inverse transformation method to reproduce the function Sample when replace=TRUE
}
\examples{
\dontrun{
set.seed(1234)
x1<-mysample(1:3,1000,c(0.2,0.3,0.5))
x1<-as.vector(table(x1))
x1/sum(x1)/c(0.2,0.3,0.5)
}
}
