% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{Rwalk}
\alias{Rwalk}
\title{A random walk Metropolis sampler}
\usage{
Rwalk(sigma)
}
\arguments{
\item{sigma}{the sd of the proposal distribution which is normal distribution}
}
\value{
a MC chain
}
\description{
A random walk Metropolis sampler for generating the standard Laplace distribution
}
\examples{
\dontrun{
set.seed(1234)
sigma<-c(0.05,0.5,5,50)
result<-lapply(sigma,Rwalk)
}
}
