\name{SA23204162-package}
\alias{SA23204162-package}
\docType{package}
\title{
  The packageof final work in 'Statistical Computing' course
}
\description{
  This package provides functions and examples of final work in 'Statistical Computing' course
}
\author{
Jingyu Huang 
Maintainer: Jingyu Huang <huangjy335@ustc.edu.cn>
}
\keyword{ package }