% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204162R.R
\name{solv}
\alias{solv}
\title{Get a inverse of matrix by R}
\usage{
solv(mat)
}
\arguments{
\item{mat}{a matrix has inverse}
}
\value{
the inverse matrix
}
\description{
inverse of matrix
}
\examples{
\dontrun{
solv(mat)
}
}
